+++ app/login/auth/admin/habitatsManager/page.tsx
@@ -1,23 +1,21 @@
-"use client"
+"use client";
 import React, { useState, useEffect } from 'react';
+import { ref, deleteObject } from 'firebase/storage';
+import { storage } from '@/lib/firebaseConfig';
 import FormCreate from '@/components/api/habitats/FormCreate';
-import FormUpdate from '@/components/api/habitats/FormUpdate'; // Importer le composant FormUpdate
+import FormUpdate from '@/components/api/habitats/FormUpdate';
 import Habitat from '@/models/habitat';
+import NekoToast from '@/components/ui/_partial/Toast';
+import { MdDelete, MdEdit } from 'react-icons/md';
 
 const HabitatsManager: React.FC = () => {
     const [habitats, setHabitats] = useState<Habitat[]>([]);
     const [loading, setLoading] = useState<boolean>(true);
     const [error, setError] = useState<string | null>(null);
-    const [showForm, setShowForm] = useState<boolean>(false); // State to control form visibility
-    const [selectedHabitat, setSelectedHabitat] = useState<Habitat | null>(null); // State to hold selected habitat for update
+    const [showForm, setShowForm] = useState<boolean>(false); 
+    const [selectedHabitat, setSelectedHabitat] = useState<Habitat | null>(null);
+    const [toast, setToast] = useState<{ type: 'Success' | 'Error' | 'Delete' | 'Update'; message: string } | null>(null);
 
-    // Function to handle successful creation of habitat
-    const handleCreateSuccess = async () => {
-        await fetchHabitats("true"); // Refresh habitats list after creation
-        setShowForm(false); // Close the create form after successful creation
-    };
-
-    // Function to fetch habitats from API
     const fetchHabitats = async (additionalParam: string | number) => {
         try {
             const response = await fetch(`/api/habitats/read?additionalParam=${encodeURIComponent(additionalParam.toString())}`);
@@ -35,27 +33,47 @@ const HabitatsManager: React.FC = () => {
         }
     };
 
-    // Function to open the create form
     const openCreateForm = () => {
         setShowForm(true);
     };
 
-    // Function to open the update form
     const openUpdateForm = (habitat: Habitat) => {
         setSelectedHabitat(habitat);
     };
 
-    // Function to close the update form
+    const handleCreateSuccess = async () => {
+        await fetchHabitats("true"); 
+        setShowForm(false); 
+        showToast('Success', 'Habitat crée avec succès');
+    };
+
     const closeUpdateForm = () => {
         setSelectedHabitat(null);
     };
 
-    // Fetch habitats when component mounts
     useEffect(() => {
         fetchHabitats('true');
     }, []);
 
     const handleDeleteHabitat = async (id: number) => {
+        const habitatToDelete = habitats.find(habitat => habitat.id === id);
+        const habitatToDeleteUrl = habitatToDelete?.imageUrl;
+        if (!habitatToDelete) {
+            console.error('Habitat not found for deletion:', id);
+            return;
+        }
+
+        try {
+            if (habitatToDeleteUrl && habitatToDeleteUrl.length > 0) {
+                for (const imageUrl of habitatToDeleteUrl) {
+                    await deleteImageFromStorage(imageUrl);
+                }
+            }
+        } catch (error) {
+            console.error('Error deleting images from storage:', error);
+            return; 
+        }
+
         try {
             const response = await fetch(`/api/habitats/delete?id=${encodeURIComponent(id.toString())}`, {
                 method: 'DELETE',
@@ -63,12 +81,31 @@ const HabitatsManager: React.FC = () => {
             const data = await response.json();
             if (data.success) {
                 setHabitats(habitats.filter(habitat => habitat.id !== id));
+                showToast('Delete', 'Habitat effacé avec succès');
             } else {
                 console.error('Error deleting habitat:', data.message);
             }
         } catch (error) {
             console.error('Error deleting habitat:', error);
         }
+    };
+
+    const deleteImageFromStorage = async (imageUrlToDelete: string) => {
+        try {
+            const storageRef = ref(storage, imageUrlToDelete);
+            await deleteObject(storageRef);
+            console.log('Successfully deleted image from storage:', imageUrlToDelete);
+        } catch (error) {
+            console.error('Error deleting image from storage:', error);
+            throw new Error('Failed to delete image from storage');
+        }
+    };
+
+    const showToast = (type: 'Success' | 'Error' | 'Delete' | 'Update', message: string) => {
+        setToast({ type, message });
+        setTimeout(() => {
+          setToast(null);
+        }, 3000);
     };
 
     if (loading) {
@@ -80,39 +117,44 @@ const HabitatsManager: React.FC = () => {
     }
 
     return (
-        <main className='flex flex-col items-center p-12'>
+        <main className='flex flex-col items-center py-12'>
+            {toast && <NekoToast toastType={toast.type} toastMessage={toast.message} timeSecond={3} onClose={() => setToast(null)} />}
+
+            <button onClick={openCreateForm} className='bg-foreground hover:bg-muted-foreground hover:text-white text-secondary py-1 px-3 rounded-md mb-6'>
+                Ajouter un habitat
+            </button>
             <h1 className='text-2xl mb-4 font-bold'>Habitats Management</h1>
-            <div className="overflow-x-auto w-full max-w-4xl">
-                <table className='min-w-full bg-white text-black shadow-md rounded-lg'>
-                    <thead className='bg-gray-200'>
+            <div className="overflow-x-auto w-full flex flex-col items-center">
+                <table className='w-full md:w-2/3 shadow-md'>
+                    <thead className='bg-muted-foreground'>
                         <tr>
-                            <th className='py-3 px-6 border-b text-left'>ID</th>
-                            <th className='py-3 px-6 border-b text-left'>Name</th>
-                            <th className='py-3 px-6 border-b text-left'>Description</th>
-                            <th className='py-3 px-6 border-b text-left'>Comment</th>
-                            <th className='py-3 px-6 border-b text-left'>Action</th>
+                            <th className='py-3 px-2 text-left'>Name</th>
+                            <th className='py-3 px-2 text-left'>Description</th>
+                            <th className='py-3 px-2 text-left'>Comment</th>
+                            <th className='py-3 px-2 text-center'>Action</th>
                         </tr>
                     </thead>
                     <tbody>
                         {habitats.map((habitat) => (
-                            <tr key={habitat.id} className='even:bg-gray-100'>
-                                <td className='py-3 px-6 border-b'>{habitat.id}</td>
-                                <td className='py-3 px-6 border-b'>{habitat.name}</td>
-                                <td className='py-3 px-6 border-b'>{habitat.description}</td>
-                                <td className='py-3 px-6 border-b'>{habitat.comment}</td>
-                                <td className='py-3 px-6 border-b text-center flex justify-center'>
-                                    <button
-                                        className='bg-red-500 hover:bg-red-600 text-white py-1 px-3 rounded mx-1 w-1/2'
-                                        onClick={() => handleDeleteHabitat(habitat.id)}
-                                    >
-                                        Supprimer
-                                    </button>
-                                    <button
-                                        className='bg-yellow-500 hover:bg-yellow-600 text-white py-1 px-3 rounded mx-1 w-1/2'
-                                        onClick={() => openUpdateForm(habitat)}
-                                    >
-                                        Modifier
-                                    </button>
+                            <tr key={habitat.id} className='bg-foreground text-secondary hover:bg-muted'>
+                                <td className='py-3 px-2 border-b-2 border-background'>{habitat.name}</td>
+                                <td className='py-3 px-2 border-b-2 border-background'>{habitat.description}</td>
+                                <td className='py-3 px-2 border-b-2 border-background'>{habitat.comment}</td>
+                                <td className='py-3 px-2 border-b-2 border-background'>
+                                    <div className='flex justify-center gap-4'>
+                                        <button
+                                            onClick={() => openUpdateForm(habitat)}
+                                            className="text-yellow-500 hover:text-yellow-600 text-[24px] md:text-[36px]"
+                                        >
+                                            <MdEdit  />
+                                        </button>
+                                        <button
+                                            onClick={() => handleDeleteHabitat(habitat.id)}
+                                            className="text-red-500 hover:text-red-600 text-[24px] md:text-[36px]"
+                                        >
+                                            <MdDelete  />
+                                        </button>
+                                    </div>
                                 </td>
                             </tr>
                         ))}
@@ -120,14 +162,6 @@ const HabitatsManager: React.FC = () => {
                 </table>
             </div>
 
-            <button
-                onClick={openCreateForm}
-                className='bg-green-500 hover:bg-green-600 text-white py-1 px-3 rounded-md mt-4'
-            >
-                Add Habitat
-            </button>
-
-            {/* FormCreate component */}
             {showForm && (
                 <FormCreate
                     onCreateSuccess={handleCreateSuccess}
@@ -135,13 +169,13 @@ const HabitatsManager: React.FC = () => {
                 />
             )}
 
-            {/* FormUpdate component */}
             {selectedHabitat && (
                 <FormUpdate
                     habitat={selectedHabitat}
                     onUpdateSuccess={() => {
-                        fetchHabitats('true'); // Refresh habitats list after update
-                        closeUpdateForm(); // Close the update form after successful update
+                        fetchHabitats('true'); 
+                        closeUpdateForm(); 
+                        showToast('Update', 'Habitat modifié avec succès');
                     }}
                     onClose={closeUpdateForm}
                 />
